add password to each button + call functions from post request
